backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
motor==3.3.2
pymongo==4.6.0
pyjwt==2.12.1
bcrypt==4.1.1
passlib==1.7.4
python-multipart==0.0.6
joblib==1.3.2
scikit-learn==1.5.0
xgboost==2.1.0
pandas==2.2.0
numpy==1.26.4
matplotlib==3.9.0
shap==0.45.0
weasyprint==60.2
reportlab==4.0.7
email-validator==2.1.0